etl/requirements.txt
flake8==6.1.0
python-dotenv==1.0.1
psycopg[binary]==3.1.18
pydantic-settings==2.6.1
elasticsearch==8.16.0